pyproject.toml
[project]
name = "bstroh"
version = "0.1.0"
description = "Static website infrastructure with AWS CDK"
authors = [{ name = "Brian Stroh", email = "brianjstroh@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aws-cdk-lib>=2.170.0",
  "constructs>=10.0.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "pre-commit>=4.0.0",
  "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infrastructure"]

[tool.ruff]
fix = true
target-version = "py311"
line-length = 88
indent-width = 2

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
strict = true

[[tool.mypy.overrides]]
module = ["aws_cdk.*", "constructs.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["infrastructure"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
]
